requirements.txt
requests
pandas
matplotlib
plotly
biopython
bcbio-gff
mafft
flask